mcnode refactor: split out REST api code
